Fix some wrong timestamp data. XML Generation working.

+++ CobaltMetrics/DataTypes/Generic/GenericList.cs
@@ -31,7 +31,7 @@ namespace CobaltMetrics.DataTypes.Generic
             data.AddRange(value);
 
             TimeSpan t = DateTime.UtcNow - new DateTime(1970, 1, 1);
-            this.timestamp = (int)t.TotalMilliseconds;
+            this.timestamp = (long)t.TotalMilliseconds;
 
             Metrics.AddData(this);
         }
